Rename TemplateRepository to TemplateDao
