+++ app/src/main/kotlin/com/ivanovsky/passnotes/data/repository/TemplateDao.kt
@@ -4,7 +4,7 @@ import com.ivanovsky.passnotes.data.entity.OperationResult
 import com.ivanovsky.passnotes.data.entity.Template
 import java.util.UUID
 
-interface TemplateRepository {
+interface TemplateDao {
     fun getTemplateGroupUid(): OperationResult<UUID?>
     fun getTemplates(): OperationResult<List<Template>>
     fun addTemplates(templates: List<Template>): OperationResult<Boolean>
