+++ app/src/main/kotlin/com/ivanovsky/passnotes/data/repository/keepass/keepass_java_2/KeepassJava2TemplateDao.kt
@@ -1,23 +1,25 @@
 @file:Suppress("FoldInitializerAndIfToElvis")
 
-package com.ivanovsky.passnotes.data.repository.keepass
+package com.ivanovsky.passnotes.data.repository.keepass.keepass_java_2
 
 import com.ivanovsky.passnotes.data.entity.GroupEntity
 import com.ivanovsky.passnotes.data.entity.OperationError.GENERIC_MESSAGE_GROUP_IS_ALREADY_EXIST
 import com.ivanovsky.passnotes.data.entity.OperationError.newDbError
 import com.ivanovsky.passnotes.data.entity.OperationResult
 import com.ivanovsky.passnotes.data.entity.Template
-import com.ivanovsky.passnotes.data.repository.TemplateRepository
+import com.ivanovsky.passnotes.data.repository.TemplateDao
 import com.ivanovsky.passnotes.data.repository.keepass.TemplateConst.TEMPLATE_GROUP_NAME
+import com.ivanovsky.passnotes.data.repository.keepass.TemplateNoteFactory
+import com.ivanovsky.passnotes.data.repository.keepass.TemplateParser
 import com.ivanovsky.passnotes.data.repository.keepass.dao.KeepassGroupDao
 import com.ivanovsky.passnotes.data.repository.keepass.dao.KeepassNoteDao
 import java.util.UUID
 import java.util.concurrent.atomic.AtomicReference
 
-class KeepassTemplateRepository(
+class KeepassJava2TemplateDao(
     private val groupDao: KeepassGroupDao,
     private val noteDao: KeepassNoteDao
-) : TemplateRepository {
+) : TemplateDao {
 
     private val templateGroupUidRef = AtomicReference<UUID>()
     private val templatesRef = AtomicReference<List<Template>>(emptyList())
